refactor: 컴포넌트명과 함수명 통일
- ApplicantCard → ApplicantCardBasic으로 변경
- ApplicantFile의 함수명을 ApplicantFileCard로 통일
- InterviewHeader 함수명 수정 (DocumentHeader → InterviewHeader)
- 파일명과 컴포넌트명 일치성 개선으로 코드 가독성 향상
